locust-simulations/requirements.txt
locust>=2.15.1
PyYAML>=6.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
matplotlib>=3.7.0
pandas>=2.0.0 